refactor(errors): remove errors store from option
Let the module to handle that itself and returns it.
