+++ frameworks/errors/ts/index.ts
@@ -1,2 +1,2 @@
 export * from './errors'
-export * from './start'
+export * from './module'
