+++ frameworks/errors/ts/module.spec.ts
@@ -1,13 +1,10 @@
-import { createStore } from '@just-web/states'
-import { ModuleError } from 'iso-error'
+import { start } from './module'
 import * as onerror from './onerror'
-import { start } from './start'
 
 jest.mock('./onerror')
 
 test('by default browser error prevent default is true', async () => {
-  const errors = createStore<ModuleError[]>([])
-  await start({ errors })
+  const { errors } = await start({})
 
   expect(onerror.registerOnErrorHandler).toBeCalledWith({
     errors,
