+++ frameworks/errors/ts/module.ts
@@ -1,10 +1,9 @@
-import { Store } from '@just-web/states'
+import { createStore } from '@just-web/states'
 import { ModuleError } from 'iso-error'
 import { registerOnErrorHandler } from './onerror'
 
 export namespace start {
   export interface Options {
-    errors: Store<ModuleError[]>,
     browserErrors?: {
       preventDefault: boolean
     }
@@ -12,8 +11,10 @@ export namespace start {
 }
 
 export async function start(options: start.Options) {
+  const errors = createStore<ModuleError[]>([])
   registerOnErrorHandler({
-    errors: options.errors,
+    errors,
     preventDefault: options.browserErrors?.preventDefault ?? true
   })
+  return { errors }
 }
